feat(backend): add neccessary DTOs for GET method endpoints

+++ backend/amoz/src/main/java/com/zpi/amoz/models/Category.java
@@ -47,7 +47,7 @@ public class Category {
         this.name = name;
     }
 
-    public int getCategoryLevel() {
+    public short getCategoryLevel() {
         return categoryLevel;
     }
